requirements.txt
feedparser==6.0.2
python-dotenv==0.14.0
python-pushover==0.4
